Make controls more intuitive
This changes the controls to be more similar to what you would see in most video games, i.e. the up-vector always points in the same direction, here in the +Y direction, and arrow keys only move in the xz-plane, not up and down. This also necessitates the introduction of dedicated up and down controls, here using the spacebar and shift keys correspondingly.